No need for UserProfilePresenterHelper anymore.
