+++ features/userprofile/impl/src/main/kotlin/io/element/android/features/userprofile/impl/root/UserProfilePresenter.kt
@@ -10,8 +10,10 @@ package io.element.android.features.userprofile.impl.root
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
@@ -22,7 +24,6 @@ import io.element.android.features.createroom.api.StartDMAction
 import io.element.android.features.userprofile.api.UserProfileEvents
 import io.element.android.features.userprofile.api.UserProfileState
 import io.element.android.features.userprofile.api.UserProfileState.ConfirmationDialog
-import io.element.android.features.userprofile.shared.UserProfilePresenterHelper
 import io.element.android.libraries.architecture.AsyncAction
 import io.element.android.libraries.architecture.AsyncData
 import io.element.android.libraries.architecture.Presenter
@@ -31,6 +32,7 @@ import io.element.android.libraries.matrix.api.MatrixClient
 import io.element.android.libraries.matrix.api.core.RoomId
 import io.element.android.libraries.matrix.api.core.UserId
 import io.element.android.libraries.matrix.api.user.MatrixUser
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
@@ -47,10 +49,23 @@ class UserProfilePresenter @AssistedInject constructor(
         fun create(userId: UserId): UserProfilePresenter
     }
 
-    private val userProfilePresenterHelper = UserProfilePresenterHelper(
-        userId = userId,
-        client = client,
-    )
+    @Composable
+    fun getDmRoomId(): State<RoomId?> {
+        return produceState<RoomId?>(initialValue = null) {
+            value = client.findDM(userId)
+        }
+    }
+
+    @Composable
+    fun getCanCall(roomId: RoomId?): State<Boolean> {
+        return produceState(initialValue = false, roomId) {
+            value = if (client.isMe(userId)) {
+                false
+            } else {
+                roomId?.let { client.getRoom(it)?.canUserJoinCall(client.sessionId)?.getOrNull() == true }.orFalse()
+            }
+        }
+    }
 
     @Composable
     override fun present(): UserProfileState {
@@ -60,8 +75,8 @@ class UserProfilePresenter @AssistedInject constructor(
         var userProfile by remember { mutableStateOf<MatrixUser?>(null) }
         val startDmActionState: MutableState<AsyncAction<RoomId>> = remember { mutableStateOf(AsyncAction.Uninitialized) }
         val isBlocked: MutableState<AsyncData<Boolean>> = remember { mutableStateOf(AsyncData.Uninitialized) }
-        val dmRoomId by userProfilePresenterHelper.getDmRoomId()
-        val canCall by userProfilePresenterHelper.getCanCall(dmRoomId)
+        val dmRoomId by getDmRoomId()
+        val canCall by getCanCall(dmRoomId)
         LaunchedEffect(Unit) {
             client.ignoredUsersFlow
                 .map { ignoredUsers -> userId in ignoredUsers }
@@ -80,7 +95,7 @@ class UserProfilePresenter @AssistedInject constructor(
                         confirmationDialog = ConfirmationDialog.Block
                     } else {
                         confirmationDialog = null
-                        userProfilePresenterHelper.blockUser(coroutineScope, isBlocked)
+                        coroutineScope.blockUser(isBlocked)
                     }
                 }
                 is UserProfileEvents.UnblockUser -> {
@@ -88,7 +103,7 @@ class UserProfilePresenter @AssistedInject constructor(
                         confirmationDialog = ConfirmationDialog.Unblock
                     } else {
                         confirmationDialog = null
-                        userProfilePresenterHelper.unblockUser(coroutineScope, isBlocked)
+                        coroutineScope.unblockUser(isBlocked)
                     }
                 }
                 UserProfileEvents.ClearConfirmationDialog -> confirmationDialog = null
@@ -118,5 +133,27 @@ class UserProfilePresenter @AssistedInject constructor(
             canCall = canCall,
             eventSink = ::handleEvents
         )
+    }
+
+    private fun CoroutineScope.blockUser(
+        isBlockedState: MutableState<AsyncData<Boolean>>,
+    ) = launch {
+        isBlockedState.value = AsyncData.Loading(false)
+        client.ignoreUser(userId)
+            .onFailure {
+                isBlockedState.value = AsyncData.Failure(it, false)
+            }
+        // Note: on success, ignoredUsersFlow will emit new item.
+    }
+
+    private fun CoroutineScope.unblockUser(
+        isBlockedState: MutableState<AsyncData<Boolean>>,
+    ) = launch {
+        isBlockedState.value = AsyncData.Loading(true)
+        client.unignoreUser(userId)
+            .onFailure {
+                isBlockedState.value = AsyncData.Failure(it, true)
+            }
+        // Note: on success, ignoredUsersFlow will emit new item.
     }
 }
